Rename project folder to "config"
This is to prepare for a single app named "tilavarauspalvelu"
