+++ .github/workflows/test.yml
@@ -110,7 +110,7 @@ jobs:
           DATABASE_URL: postgis://tvp:tvp@localhost:5432/tvp
 
       - name: "Test for missing translation"
-        run: poetry run python -m tilavarauspalvelu.hooks.translations_done
+        run: poetry run python -m config.hooks.translations_done
         env:
           DJANGO_SETTINGS_ENVIRONMENT: CI
           DATABASE_URL: postgis://tvp:tvp@localhost:5432/tvp
