+++ api/helauth/urls.py
@@ -1,5 +1,5 @@
 # This file exists so that we can use a custom `TunnistamoOIDCAuth` backend.
-# See `tilavarauspalvelu.auth.ProxyTunnistamoOIDCAuthBackend` for more details.
+# See `config.auth.ProxyTunnistamoOIDCAuthBackend` for more details.
 # Using `helusers/urls.py` will not add the URLs to the file's `urlpatterns` unless
 # `helusers.tunnistamo_oidc.TunnistamoOIDCAuth` is specifically included in the
 # `AUTHENTICATION_BACKENDS` setting.
