+++ config/__init__.py
@@ -1,5 +1,5 @@
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
-from tilavarauspalvelu.celery import app as celery_app
+from config.celery import app as celery_app
 
 __all__ = ("celery_app",)
