+++ config/settings.py
@@ -35,8 +35,8 @@ class Common(Environment):
 
     # --- Basic settings ---------------------------------------------------------------------------------------------
 
-    WSGI_APPLICATION = "tilavarauspalvelu.wsgi.application"
-    ROOT_URLCONF = "tilavarauspalvelu.urls"
+    WSGI_APPLICATION = "config.wsgi.application"
+    ROOT_URLCONF = "config.urls"
     AUTH_USER_MODEL = "users.User"
     DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
     SECRET_KEY = values.StringValue()
@@ -97,15 +97,15 @@ class Common(Environment):
     ]
 
     MIDDLEWARE = [
-        "tilavarauspalvelu.middleware.QueryLoggingMiddleware",
-        "tilavarauspalvelu.middleware.MultipleProxyMiddleware",
+        "config.middleware.QueryLoggingMiddleware",
+        "config.middleware.MultipleProxyMiddleware",
         "corsheaders.middleware.CorsMiddleware",
         "django.middleware.security.SecurityMiddleware",
         # Keep this after security middleware, correct place according to whitenoise documentation
         "whitenoise.middleware.WhiteNoiseMiddleware",
         "django.contrib.sessions.middleware.SessionMiddleware",
         # Must be after the session middleware, since it looks for a flag in the session
-        "tilavarauspalvelu.middleware.KeycloakRefreshTokenExpiredMiddleware",
+        "config.middleware.KeycloakRefreshTokenExpiredMiddleware",
         "django.middleware.common.CommonMiddleware",
         "django.middleware.csrf.CsrfViewMiddleware",
         "django.contrib.auth.middleware.AuthenticationMiddleware",
@@ -232,7 +232,7 @@ class Common(Environment):
             "filters": {},
             "formatters": {
                 "common": {
-                    "()": "tilavarauspalvelu.logging.TVPFormatter",
+                    "()": "config.logging.TVPFormatter",
                     "format": (
                         "Time: {asctime} "
                         "| Level: {levelname} "
@@ -287,8 +287,8 @@ class Common(Environment):
     # --- Authentication settings ------------------------------------------------------------------------------------
 
     AUTHENTICATION_BACKENDS = [
-        "tilavarauspalvelu.auth.ProxyTunnistamoOIDCAuthBackend",
-        "tilavarauspalvelu.auth.ProxyModelBackend",
+        "config.auth.ProxyTunnistamoOIDCAuthBackend",
+        "config.auth.ProxyModelBackend",
     ]
 
     AUTH_PASSWORD_VALIDATORS = [
@@ -393,7 +393,7 @@ class Common(Environment):
     GRAPHENE = {
         "SCHEMA": "api.graphql.schema.schema",
         "MIDDLEWARE": [
-            "tilavarauspalvelu.middleware.GraphQLSentryMiddleware",
+            "config.middleware.GraphQLSentryMiddleware",
         ],
     }
 
@@ -632,7 +632,7 @@ class Local(Common, overrides_from=LocalMixin):
         "SCHEMA": Common.GRAPHENE["SCHEMA"],
         "MIDDLEWARE": [
             "graphene_django.debug.DjangoDebugMiddleware",
-            "tilavarauspalvelu.middleware.GraphQLErrorLoggingMiddleware",
+            "config.middleware.GraphQLErrorLoggingMiddleware",
         ],
     }
 
@@ -756,7 +756,7 @@ class AutomatedTests(EmptyDefaults, Common, dotenv_path=None, overrides_from=Aut
         "SCHEMA": Common.GRAPHENE["SCHEMA"],
         "TESTING_ENDPOINT": "/graphql/",
         "MIDDLEWARE": [
-            "tilavarauspalvelu.middleware.GraphQLErrorLoggingMiddleware",
+            "config.middleware.GraphQLErrorLoggingMiddleware",
         ],
     }
