+++ config/urls.py
@@ -14,7 +14,7 @@ original_each_context = admin.site.each_context
 admin.site.each_context = lambda request: original_each_context(request) | {
     "version": settings.APP_VERSION,
     # The helauth variables need to be added, since we subclass `helusers.tunnistamo_oidc.TunnistamoOIDCAuth`
-    # with `tilavarauspalvelu.auth.ProxyTunnistamoOIDCAuthBackend` for optimizing request user fetching.
+    # with `config.auth.ProxyTunnistamoOIDCAuthBackend` for optimizing request user fetching.
     # `helusers.admin_site.AdminSite.each_context` refers to the original backend by
     # string reference, so subclasses won't have the login/logout urls added.
     "helsinki_provider_installed": True,
