+++ docker/uwsgi.yml
@@ -5,8 +5,8 @@ uwsgi:
   # Needed plugins if running against Debian uwsgi-package
   # python docker image cannot use that due to linker mishaps
   # plugins: python3,http
-  module: tilavarauspalvelu.wsgi:application
-  # wsgi-file: tilavarauspalvelu/wsgi.p
+  module: config.wsgi:application
+  # wsgi-file: config/wsgi.p
   #uid: tvp
   umask: 022
   http: :8000
