+++ email_notification/helpers/email_sender.py
@@ -17,10 +17,10 @@ if TYPE_CHECKING:
     from collections.abc import Iterable
 
     from applications.models import Application
+    from config.utils.commons import LanguageType
     from email_notification.admin.email_template_tester import EmailTemplateTesterForm
     from email_notification.helpers.email_builder_base import BaseEmailBuilder
     from reservations.models import Reservation
-    from tilavarauspalvelu.utils.commons import LanguageType
 
 
 class EmailNotificationSender:
