+++ permissions/models/general_role.py
@@ -6,8 +6,8 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from graphene_django_extensions.fields.model import StrChoiceField
 
+from config.utils.auditlog_util import AuditLogger
 from permissions.enums import UserRoleChoice
-from tilavarauspalvelu.utils.auditlog_util import AuditLogger
 
 if TYPE_CHECKING:
     import datetime
