+++ sonar-project.properties
@@ -9,10 +9,10 @@ sonar.python.version=3.11
 sonar.source=./
 
 # Path globs to files excluded from scanning
-sonar.exclusions=**tests**/*,**migrations**/*,**Dockerfile*,**docker-compose*,tilavarauspalvelu/settings.py
+sonar.exclusions=**tests**/*,**migrations**/*,**Dockerfile*,**docker-compose*,config/settings.py
 
 # Path globs to files excluded from coverage
-sonar.coverage.exclusions=**tests**/*,**migrations**/*,**admin**/*,tilavarauspalvelu/settings.py,manage.py,local_settings_example.py
+sonar.coverage.exclusions=**tests**/*,**migrations**/*,**admin**/*,config/settings.py,manage.py,local_settings_example.py
 
 # Path to coverage file generated in GitHub Actions
 sonar.python.coverage.reportPaths=coverage.xml
