+++ tests/test_email/test_email_context_application.py
@@ -2,8 +2,8 @@ import datetime
 
 import pytest
 
+from config.utils.commons import LanguageType
 from email_notification.helpers.email_builder_application import ApplicationEmailContext
-from tilavarauspalvelu.utils.commons import LanguageType
 
 
 @pytest.mark.parametrize("language", ["fi", "en", "sv"])
